+++ src/api/userLogin.js
@@ -1,7 +1,7 @@
 import meet from "./checkAuthorized.js";
 import { USER_LOGIN } from "./config";
 
-export const userLogin = async (loginID, password) => {
+export const handleUserLogin = async (loginID, password) => {
   try {
     const response = await meet.post(
       USER_LOGIN,
@@ -17,7 +17,15 @@ export const userLogin = async (loginID, password) => {
 
     return { user_id, username };
   } catch (error) {
-    console.error("Error during user login:", error);
-    throw error;
+    if (error.response.status === 401) {
+      console.error("잘못된 아이디 또는 비밀번호를 입력했습니다.");
+      alert("잘못된 아이디 또는 비밀번호를 입력했습니다.");
+      return null;
+    }
+    console.error("로그인 요청 중 에러가 발생했습니다.:", error);
+    alert(
+      `로그인 요청 중 에러가 발생했습니다. 관리자에게 문의하세요\nError Code: ${error.response.status}\nMessage: ${error.message}`
+    );
+    return null;
   }
 };
